Module arceos_api::config

Platform-specific constants and parameters.

Constants

ARCH

Architecture identifier.

FAMILY

Platform family.

KERNEL_ASPACE_BASE

Kernel address space base.

KERNEL_ASPACE_SIZE

Kernel address space size.

KERNEL_BASE_PADDR

Base physical address of the kernel image.

KERNEL_BASE_VADDR

Base virtual address of the kernel image.

MMIO_REGIONS

MMIO regions with format (base_paddr, size).

PCI_BUS_END

End PCI bus number.

PCI_ECAM_BASE

Base physical address of the PCIe ECAM space (should read from ACPI ‘MCFG’ table).

PCI_RANGES

PCI device memory ranges (not used on x86).

PHYS_BUS_OFFSET

Offset of bus address and phys address. some boards, the bus address is different from the physical address.

PHYS_MEMORY_BASE

Base address of the whole physical memory.

PHYS_MEMORY_END

End address of the whole physical memory.

PHYS_MEMORY_SIZE

Size of the whole physical memory.

PHYS_VIRT_OFFSET

Linear mapping offset, for quick conversions between physical and virtual addresses.

PLATFORM

Platform identifier.

SMP

Number of CPUs

TASK_STACK_SIZE

Stack size of each task.

TICKS_PER_SEC

Number of timer ticks per second (Hz). A timer tick may contain several timer interrupts.

TIMER_FREQUENCY

Timer interrupt frequencyin Hz.

VIRTIO_MMIO_REGIONS

VirtIO MMIO regions with format (base_paddr, size).