arceos_api

Enum AxError

#[non_exhaustive]

#[repr(i32)]
pub enum AxError {

    AddrInUse = 1,
    AlreadyConnected = 2,
    AlreadyExists = 3,
    ArgumentListTooLong = 4,
    BadAddress = 5,
    BadFileDescriptor = 6,
    BadState = 7,
    BrokenPipe = 8,
    ConnectionRefused = 9,
    ConnectionReset = 10,
    CrossesDevices = 11,
    DirectoryNotEmpty = 12,
    FilesystemLoop = 13,
    IllegalBytes = 14,
    InProgress = 15,
    Interrupted = 16,
    InvalidData = 17,
    InvalidExecutable = 18,
    InvalidInput = 19,
    Io = 20,
    IsADirectory = 21,
    NameTooLong = 22,
    NoMemory = 23,
    NoSuchDevice = 24,
    NoSuchProcess = 25,
    NotADirectory = 26,
    NotASocket = 27,
    NotATty = 28,
    NotConnected = 29,
    NotFound = 30,
    OperationNotPermitted = 31,
    OperationNotSupported = 32,
    OutOfRange = 33,
    PermissionDenied = 34,
    ReadOnlyFilesystem = 35,
    ResourceBusy = 36,
    StorageFull = 37,
    TimedOut = 38,
    TooManyOpenFiles = 39,
    UnexpectedEof = 40,
    Unsupported = 41,
    WouldBlock = 42,
    WriteZero = 43,

}

The error kind type used by ArceOS.

Similar to std::io::ErrorKind.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AddrInUse = 1

A socket address could not be bound because the address is already in use elsewhere.

AlreadyConnected = 2

The socket is already connected.

AlreadyExists = 3

An entity already exists, often a file.

ArgumentListTooLong = 4

Program argument list too long.

BadAddress = 5

Bad address.

BadFileDescriptor = 6

Bad file descriptor.

BadState = 7

Bad internal state.

BrokenPipe = 8

Broken pipe

ConnectionRefused = 9

The connection was refused by the remote server.

ConnectionReset = 10

The connection was reset by the remote server.

CrossesDevices = 11

Cross-device or cross-filesystem (hard) link or rename.

DirectoryNotEmpty = 12

A non-empty directory was specified where an empty directory was expected.

FilesystemLoop = 13

Loop in the filesystem or IO subsystem; often, too many levels of symbolic links.

IllegalBytes = 14

Illegal byte sequence.

InProgress = 15

The operation was partially successful and needs to be checked later on due to not blocking.

Interrupted = 16

This operation was interrupted.

InvalidData = 17

Data not valid for the operation were encountered.

Unlike InvalidInput, this typically means that the operation parameters were valid, however the error was caused by malformed input data.

For example, a function that reads a file into a string will error with InvalidData if the file’s contents are not valid UTF-8.

InvalidExecutable = 18

Invalid executable format.

InvalidInput = 19

Invalid parameter/argument.

Io = 20

Input/output error.

IsADirectory = 21

The filesystem object is, unexpectedly, a directory.

NameTooLong = 22

Filename is too long.

NoMemory = 23

Not enough space/cannot allocate memory.

NoSuchDevice = 24

No such device.

NoSuchProcess = 25

No such process.

NotADirectory = 26

A filesystem object is, unexpectedly, not a directory.

NotASocket = 27

The specified entity is not a socket.

NotATty = 28

Not a typewriter.

NotConnected = 29

The network operation failed because it was not connected yet.

NotFound = 30

The requested entity is not found.

OperationNotPermitted = 31

Operation not permitted.

OperationNotSupported = 32

Operation not supported.

OutOfRange = 33

Result out of range.

PermissionDenied = 34

The operation lacked the necessary privileges to complete.

ReadOnlyFilesystem = 35

The filesystem or storage medium is read-only, but a write operation was attempted.

ResourceBusy = 36

Device or resource is busy.

StorageFull = 37

The underlying storage (typically, a filesystem) is full.

TimedOut = 38

The I/O operation’s timeout expired, causing it to be canceled.

TooManyOpenFiles = 39

The process has too many files open.

UnexpectedEof = 40

An error returned when an operation could not be completed because an “end of file” was reached prematurely.

Unsupported = 41

This operation is unsupported or unimplemented.

WouldBlock = 42

The operation needs to block to complete, but the blocking operation was requested to not occur.

WriteZero = 43

An error returned when an operation could not be completed because a call to write() returned Ok(0).

Implementations

impl AxErrorKind

pub fn as_str(&self) -> &'static str

Returns the error description.

pub const fn code(self) -> i32

Returns the error code value in i32.