arceos_api::fs

Function ax_read_dir

pub fn ax_read_dir(
    dir: &mut AxDirHandle,
    dirents: &mut [AxDirEntry],
) -> AxResult<usize>

Available on crate feature fs only.

Reads directory entries starts from the current position into the given buffer, returns the number of entries read.

After the read, the cursor of the directory will be advanced by the number of entries read.