arceos_api

Module mem

Memory management.

Structs

DMAInfodma

Represents information related to a DMA operation.

Functions

ax_alloc⚠alloc

Allocates a continuous memory blocks with the given layout in the global allocator.

ax_alloc_coherent⚠dma

Allocates coherent memory that meets Direct Memory Access (DMA) requirements.

ax_dealloc⚠alloc

Deallocates the memory block at the given ptr pointer with the given layout, which should be allocated by ax_alloc.

ax_dealloc_coherent⚠dma

Deallocates coherent memory previously allocated.