arceos_api
Module net

Networking primitives for TCP/UDP communication.

Structs

AxTcpSocketHandlenet

A handle to a TCP socket.

AxUdpSocketHandlenet

A handle to a UDP socket.

Functions

ax_dns_querynet

Resolves the host name to a list of IP addresses.

ax_poll_interfacesnet

Poll the network stack.

ax_tcp_acceptnet

Accepts a new connection on the TCP socket.

ax_tcp_bindnet

Binds the TCP socket to the given address and port.

ax_tcp_connectnet

Connects the TCP socket to the given address and port.

ax_tcp_listennet

Starts listening on the bound address and port.

ax_tcp_peer_addrnet

Returns the remote address and port of the TCP socket.

ax_tcp_pollnet

Returns whether the TCP socket is readable or writable.

ax_tcp_recvnet

Receives data on the TCP socket, and stores it in the given buffer. On success, returns the number of bytes read.

ax_tcp_sendnet

Transmits data in the given buffer on the TCP socket.

ax_tcp_set_nonblockingnet

Moves this TCP socket into or out of nonblocking mode.

ax_tcp_shutdownnet

Closes the connection on the TCP socket.

ax_tcp_socketnet

Creates a new TCP socket.

ax_tcp_socket_addrnet

Returns the local address and port of the TCP socket.

ax_udp_bindnet

Binds the UDP socket to the given address and port.

ax_udp_connectnet

Connects this UDP socket to a remote address, allowing the send and recv to be used to send data and also applies filters to only receive data from the specified address.

ax_udp_peek_fromnet

Receives a single datagram message on the UDP socket, without removing it from the queue.

ax_udp_peer_addrnet

Returns the remote address and port of the UDP socket.

ax_udp_pollnet

Returns whether the UDP socket is readable or writable.

ax_udp_recvnet

Receives a single datagram message on the UDP socket from the remote address to which it is connected. On success, returns the number of bytes read.

ax_udp_recv_fromnet

Receives a single datagram message on the UDP socket.

ax_udp_sendnet

Sends data on the UDP socket to the remote address to which it is connected.

ax_udp_send_tonet

Sends data on the UDP socket to the given address. On success, returns the number of bytes written.

ax_udp_set_nonblockingnet

Moves this UDP socket into or out of nonblocking mode.

ax_udp_socketnet

Creates a new UDP socket.

ax_udp_socket_addrnet

Returns the local address and port of the UDP socket.