Module arceos_api::stdio

Standard input and output.

Functions

ax_console_read_byte

Reads a byte from the console, or returns None if no input is available.

ax_console_write_bytes

Writes a slice of bytes to the console, returns the number of bytes written.

ax_console_write_fmt

Writes a formatted string to the console.