arceos_api

Module fs

Filesystem manipulation operations.

Structs

AxDirHandlefs

A handle to an opened directory.

AxDiskfs and myfs

A disk device with a cursor.

AxFileHandlefs

A handle to an opened file.

AxOpenOptionsfs

Options and flags which can be used to configure how a file is opened.

Enums

AxSeekFromfs

Enumeration of possible methods to seek within an I/O object.

Traits

MyFileSystemIffs and myfs

The interface to define custom filesystems in user apps.

Functions

ax_create_dirfs

Creates a new, empty directory at the provided path.

ax_current_dirfs

Returns the current working directory.

ax_file_attrfs

Returns attributes of the file.

ax_flush_filefs

Flushes the file, writes all buffered data to the underlying device.

ax_open_dirfs

Opens a directory at the path relative to the current directory with the options specified by opts.

ax_open_filefs

Opens a file at the path relative to the current directory with the options specified by opts.

ax_read_dirfs

Reads directory entries starts from the current position into the given buffer, returns the number of entries read.

ax_read_filefs

Reads the file at the current position, returns the number of bytes read.

ax_read_file_atfs

Reads the file at the given position, returns the number of bytes read.

ax_remove_dirfs

Removes an empty directory.

ax_remove_filefs

Removes a file from the filesystem.

ax_renamefs

Rename a file or directory to a new name.

ax_seek_filefs

Sets the cursor of the file to the specified offset. Returns the new position after the seek.

ax_set_current_dirfs

Changes the current working directory to the specified path.

ax_truncate_filefs

Truncates the file to the specified size.

ax_write_filefs

Writes the file at the current position, returns the number of bytes written.

ax_write_file_atfs

Writes the file at the given position, returns the number of bytes written.

Type Aliases

AxDirEntryfs

Alias of [axfs_vfs::VfsDirEntry].

AxFileAttrfs

Alias of [axfs_vfs::VfsNodeAttr].

AxFilePermfs

Alias of [axfs_vfs::VfsNodePerm].

AxFileTypefs

Alias of [axfs_vfs::VfsNodeType].